feat: remove no check option
