+++ README.md
@@ -45,8 +45,6 @@ yarn add quickbundle
 	"scripts": {
 		"build": "quickbundle build", // Production mode (optimizes bundle)
 		"watch": "quickbundle watch", // Development mode (watches each file change)
-		"build:fast": "quickbundle build --no-check", // Production mode with fast transpilation mode. This mode disables TypeScript type checking (ie. not using `tsc`) and, consequently, the `types` is no more generated by the tool
-		"watch:fast": "quickbundle watch --no-check" // Development mode with fast transpilation mode
 	}
 }
 ```
