--- examples/module/package.json
@@ -1,19 +1,0 @@
-{
-	"name": "vanilla-module",
-	"private": true,
-	"source": "src/index.js",
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.mjs",
-	"scripts": {
-		"clean": "rm -rf dist",
-		"build": "quickbundle build",
-		"watch": "quickbundle watch"
-	},
-	"devDependencies": {
-		"quickbundle": "link:../../",
-		"preact": "10.13.2"
-	},
-	"peerDependencies": {
-		"preact": ">=10.0.0"
-	}
-}
