--- examples/module/src/Button.jsx
@@ -1,7 +1,0 @@
-export const Button = () => {
-	return (
-		<>
-			<button>Click me!</button>Plop
-		</>
-	);
-};
