--- examples/preact/package.json
@@ -1,17 +1,0 @@
-{
-	"name": "vanilla-preact",
-	"private": true,
-	"source": "src/index.jsx",
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.mjs",
-	"scripts": {
-		"clean": "rm -rf dist",
-		"start": "pnpm watch",
-		"build": "quickbundle build",
-		"watch": "quickbundle watch"
-	},
-	"devDependencies": {
-		"quickbundle": "link:../../",
-		"preact": "10.13.2"
-	}
-}
