--- examples/ts-module/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "ts-module",
-	"private": true,
-	"source": "src/index.ts",
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.mjs",
-	"types": "dist/index.d.ts",
-	"scripts": {
-		"clean": "rm -rf dist",
-		"build": "quickbundle build",
-		"watch": "quickbundle watch",
-		"build:fast": "quickbundle build --no-check",
-		"watch:fast": "quickbundle watch --no-check"
-	},
-	"devDependencies": {
-		"quickbundle": "link:../../",
-		"preact": "10.13.2",
-		"typescript": "5.0.3"
-	},
-	"peerDependencies": {
-		"preact": ">=10.0.0"
-	}
-}
