--- examples/ts-module/src/Button.tsx
@@ -1,14 +1,0 @@
-interface ButtonProps {
-	label: string;
-}
-
-/**
- * Button component
- */
-export const Button = (props: ButtonProps) => {
-	return (
-		<>
-			<button>{props.label}</button> Plop
-		</>
-	);
-};
