--- examples/ts-module/src/index.ts
@@ -1,6 +1,0 @@
-export { Button } from "./Button";
-
-export const version = "0.0.0";
-
-// eslint-disable-next-line import/no-default-export
-export default "ts-module";
