--- examples/ts-preact/package.json
@@ -1,23 +1,0 @@
-{
-	"name": "ts-preact",
-	"private": true,
-	"source": "src/index.tsx",
-	"main": "dist/index.cjs.js",
-	"module": "dist/index.mjs",
-	"types": "dist/index.d.ts",
-	"scripts": {
-		"clean": "rm -rf dist",
-		"start": "pnpm watch",
-		"build": "quickbundle build",
-		"watch": "quickbundle watch",
-		"build:fast": "quickbundle build --no-check",
-		"watch:fast": "quickbundle watch --no-check"
-	},
-	"devDependencies": {
-		"typescript": "5.0.3"
-	},
-	"dependencies": {
-		"quickbundle": "link:../../",
-		"preact": "10.13.2"
-	}
-}
