--- examples/ts-preact/src/index.tsx
@@ -1,4 +1,0 @@
-import { render } from "preact";
-import { App } from "./App";
-
-render(<App />, document.getElementById("root"));
