+++ examples/ts-react/package.json
@@ -9,18 +9,16 @@
 		"clean": "rm -rf dist",
 		"start": "pnpm watch",
 		"build": "quickbundle build",
-		"watch": "quickbundle watch",
-		"build:fast": "quickbundle build --no-check",
-		"watch:fast": "quickbundle watch --no-check"
-	},
-	"devDependencies": {
-		"@types/react": "18.0.31",
-		"@types/react-dom": "18.0.11",
-		"typescript": "5.0.3"
+		"watch": "quickbundle watch"
 	},
 	"dependencies": {
 		"quickbundle": "link:../../",
 		"react": "18.2.0",
 		"react-dom": "18.2.0"
+	},
+	"devDependencies": {
+		"@types/react": "18.0.31",
+		"@types/react-dom": "18.0.11",
+		"typescript": "5.0.3"
 	}
 }
