+++ examples/ts-react/src/assets/image.gif
