+++ examples/ts-react/src/assets/image.jpeg
