+++ examples/ts-react/src/assets/image.jpg
