+++ examples/ts-react/src/assets/image.png
