+++ examples/ts-react/tsconfig.json
@@ -4,5 +4,6 @@
 		"moduleResolution": "Node",
 		"target": "ES2015"
 	},
-	"include": ["src"]
+	"include": ["src", "types"],
+	"exclude": ["node_modules"]
 }
