+++ examples/ts-react/types/module.d.ts
