--- src/program/types.ts
@@ -1,3 +1,0 @@
-export type ProgramContext = {
-	noCheck: boolean;
-};
